Update sample code to give expected output
The change in line 97 is because the 0th iteration of the `for` loop 
would print on the same line as the column headers, the newline 
vs. tab means this starts on line below the headers.

The change in line 103 is because the header "NO FORMAT" 
extends past the first indent position, causing the outputs to 
line up immediately under incorrect columns headers, and only 
with much care does one avoid the mistake of reading the 
wrong column based on the header.

The change in line 22 is to set the text as code in the display.