Explain difference between pulseIn() and pulseInLong()
Likely the question "What's the difference between this and pulseIn()" will be the first thing a user will want to know when visiting this page. Although this information was somewhat alluded to at the end of the description, it's much better for it to be at the start of the text.

Reference:
https://github.com/arduino/Arduino/pull/2495#issue-25936527