Use consistent formatting in pulseInLong() description
Code markdown was used on some instances of the words HIGH and LOW and not on others. A search of the rest of the reference pages shows them to be quite inconsistent in formatting but code markdown is used in the majority of instances and I feel that's the correct usage.