Correct recommendations re: use of Serial.print() to debug Keyboard/Mouse emulation
The recommendation in the Keyboard and Mouse reference pages said to refer to those library's reference pages for examples of how to use Serial.print() to debug the output. However, none of the examples have such code. I believe that statement was actually intended to be put on the other recommendation to use a control system as the Keyboard/Mouse reference pages do have example code for this.

As a replacement, I added a link to the Serial.print() reference page.