Add Yun to the list of the boards serialEvent() doesn't work for
I had originally left the Yun off the list because it's retired and I figure "Leonardo and Micro" can represent all ATmega32U4-based boards. However, there is an issue report specifically requesting that this board be added to the list so I'll add it in just to be able to conclusively close that issue.

I don't think it does any harm to be a bit more verbose, since this list is in the "Notes and Warnings" section.