Add unsigned int/long type casting pages
I'm not certain whether casting pages for all the types in the reference should be added, but certainly it must be demonstrated that casting can be done to the multi-word types as well. unsigned int/unsigned long are important types, so at a minimum these must be added.

Likely it would have been best to write a single generic page on type casting but we took the other fork in the road too long ago to turn back now.