Separate parameters list with blank lines
When there were multiple parameters with longer descriptions, the Parameters section ended up looking like a wall of text. Separating each parameter with a blank line makes it easier to read and faster to find the information you're searching for.

This format is also easier for the collaborators to work with because they don't need to remember to add the ` +` at the end of the line. There are currently several reference pages where that was forgotten.

I had also considered formatting the parameters section as a list. That looks quite nice also, but in the end I decided I slightly preferred the blank lines.

Example of the previous formatting being problematic:
https://www.arduino.cc/reference/en/language/functions/external-interrupts/attachinterrupt#_parameters