Establish a standard for the use of blank lines to separate sections
Blank lines were used inconsistently and unnecessarily in the reference example. I have attempted to decipher the intended spacing standard and apply this across the document.

This proposal establishes an official style. If accepted, I will apply it across all the reference pages, where this sort of inconsistency is widespread.