Use beginner-friendly code as an example for #include
The previous example code was very advanced, as well as using a type that is no longer supported. #include is a part of the language that the user will encounter fairly early on and so the documentation should be beginner-friendly. The Servo library seems like a very common library (since it's bundled with the Arduino IDE), that can be demonstrated with fairly simple, straightforward code.