Clarify documentation in attach/detachInterrupt syntax section
Feedback was recieved that it wasn't clear whether the second sentence of the attachInterrupt(pin, ISR, mode) syntax description was related to the first sentence (i.e. is the syntax not recommended for Arduino SAMD Boards, Uno WiFi Rev2, Due, 101 only and the syntax is recommended for any other board type?).