Swap and correct the documentation of isSpace() and isWhitespace()
Although the documentation previously matched what you might expect these functions to do based on their names, the actual behavior of the functions is the reverse. In addition to this, isWhitespace() matches on horizontal tab as well as space, which was not mentioned in the isSpace() documentation. The decision was made to leave the implementation of the functions as-is to avoid causing breakage and to correct the documentation to match the actual behavior of the functions.

More information:
https://github.com/arduino/ArduinoCore-API/pull/27