Update readBytesUntil.adoc
@per1234: 
Changed the warning to make it more beginners-friendly. Also changed the order of termination checks as discussed in previous pull request.

Also changed the description of what a returned 0 means.

As a new git user, I wasn't sure whether to create a new branch for this PR, hope this works for you.