Remove "Notes and Warnings" section
The zero-indexed nature of the macro is already documented in the parameters section.

The handling of zeroed bits is already clear from the description.

The requirement for the parameters to be integers seems fairly obvious to me, but I'm open to being convinced otherwise. However, if it is necessary to document the supported parameter types, that should be done in the parameters section, following the convention that has been established through the rest of the language reference content, so this part needs to be removed either way.