Update millis.adoc
changed variable named "time" to "my_time", because it was a conflict when compiling on ESP32 board; 
And it is not a good practice to define such generic variables that could be used in system variables. 
Using user variables like my_time and my_temp_sensor is much easier to understand for beginners and more intuitive, easier to differentiate from the system variables.