Update millis.adoc
I am an infrequent Arduino user, and when I come back to it, my memory needs a jog.  The existing example for the millis() function really doesn't demonstrate HOW to use it as a substitute for the Delay_ms() function -- which, in my case, is the main reason I would use millis().  At any rate, I felt this write up needed a more functional example, so I supplied one.  I fully tested this code on an Arduino UNO [but only on an UNO], in the Arduino IDE version 1.8.13, using the IDE supplied Serial Monitor.