Update attachInterrupt.adoc for MEGA
Using pins 20, 21 for interrupts on the MEGA 2560 can cause some real headaches if not aware that they have external pull up resistors on them. The pull ups are there for TWI functionality, and cannot be disabled shy of desoldering the resistors from the board. 
I can confirm that the 2560 and the ADK both have these resistors. I do not know about the original MEGA, but I assume it has them too.