Update while.adoc
None of the While or While Loop explanations or tutorials make it entirely clear how the loop ends or what happens when it does end, and I've puzzled over this for some time.  Also, some comparison is needed between If, While, Do While, and For to clarify what each is best for and when best to use them.
Also, NONE of these While explanations or tutorials mentions that you can use <break> to get out of them early.   An example using Break would be good on this page, too, such as using a sensor reading to interrupt the While before it's finished.  Let's say, while a switch is open, keep running a motor until the switch closes.  Can do this with If statement, how is while better?